Score is judged by value of enemies defeated
Instead of getting 1 point of score for each enemy defeated, different enemies can now be worth different amounts. At the end of the game the total number of points the player acquired are shown (currently with trailing zeroes).

+++ abundantadversaries/src/com/dtlinley/abundantadversaries/game/World.java
@@ -47,7 +47,7 @@ public class World implements Renderable {
 		return textures;
 	}
 
-	public int getScore() {
+	public float getScore() {
 		return director.getScore();
 	}
